refactor: move init.py to apis subpackage
